+++ src/thermohw/__init__.py
@@ -1,6 +1,6 @@
 """Convert thermo homework assignments."""
 
-from .convert_thermo_hw import nb_exp, pdf_exp, process as hw_process  # noqa: F401
+from .convert_thermo_hw import process as hw_process  # noqa: F401
 from .extract_attachments import ExtractAttachmentsPreprocessor  # noqa: F401
 from .pymarkdown import PyMarkdownPreprocessor  # noqa: F401
 from .preprocessors import RawRemover, SolutionRemover  # noqa: F401
@@ -12,10 +12,3 @@ from .filters import (  # noqa: F401
     convert_raw_html,
 )
 from ._version import __version__  # noqa: F401
-from .convert_thermo_exam import (  # noqa: F401
-    sa_nb_exp,
-    prob_nb_exp,
-    process as exam_process,
-    solution_pdf_exp as exam_pdf_exp,
-    solution_nb_exp as exam_nb_exp,
-)