+++ test/test_convert_thermo_hw.py
@@ -2,7 +2,8 @@
 import os
 import pkg_resources
 
-from ..convert_thermo_hw import pdf_exp
+import nbformat
+from thermohw.convert_thermo_hw import pdf_exp, nb_exp
 
 
 def test_convert_pathological_image_name() -> None:
@@ -16,3 +17,15 @@ def test_convert_pathological_image_name() -> None:
     }
     solution_pdf, _ = pdf_exp.from_filename(filename, res)
     assert len(solution_pdf) > 0
+
+
+def test_tags() -> None:
+    """Test tag support."""
+    filename = os.path.join("test-cell-tags.ipynb")
+    filename = pkg_resources.resource_filename(__name__, filename)
+    res = {"remove_solution": True, "legacy": False}
+    problem_nb = nbformat.read(filename, as_version=4)
+    if "celltoolbar" in problem_nb.metadata:
+        del problem_nb.metadata["celltoolbar"]
+    solution_nb, _ = nb_exp.from_notebook_node(problem_nb, res)
+    assert len(solution_nb) > 0